fixes node randomizer HR, but aborts the changes in the node randomizer to allow for more than 0 mutations along branches
fixes the HR issue in the node randomizer by conditioning on both children of the grand parent node of each node to have 0 mutations. Started working on a more elaborate version of the operator that would allow for mutations larger than 0, that work is currently in the noderandomzier2 class. Could be revisited in the future to make more elaborate changes